fix: sync Flutter app with production backend
P0 Critical:
- Change API base URL default to https://alochi.org
- Rewrite test flow: 3-step attempt/answers/submit instead of broken single submit
- Replace hardcoded mock test results with real API fetch
- Fix TestResultModel to map backend field names

P1 Major:
- Fix dashboard stats: total_xp, consecutive_days_active, separate coins fetch
- Fix token refresh to save rotated refresh token
- Remove unused variables and duplicate casts

+++ lib/features/student/shop/shop_screen.dart
@@ -220,9 +220,9 @@ class _ShopScreenState extends ConsumerState<ShopScreen> {
                           icon: Icons.storefront_outlined);
                     }
                     final balance =
-                        walletAsync.value?['coins'] ??
-                            walletAsync.value?['balance'] ??
-                            0;
+                        (walletAsync.value?['coins'] ??
+                                walletAsync.value?['balance'] ??
+                                0) as int;
                     return LayoutBuilder(
                         builder: (ctx, constraints) {
                       final cols =
@@ -239,9 +239,9 @@ class _ShopScreenState extends ConsumerState<ShopScreen> {
                         itemCount: items.length,
                         itemBuilder: (ctx, i) => _ShopItemCard(
                           item: items[i],
-                          balance: balance as int,
+                          balance: balance,
                           onPurchase: () =>
-                              _purchase(items[i], balance as int),
+                              _purchase(items[i], balance),
                         ),
                       );
                     });
